add helpers for seeding records
